Created a checkout system with transaction rules
Implemented a program in Java that reads pricing rules from
a txt file for specific products, checks to see if there any
discounts for some items, then displays the pricing information
and usage rules to the client.
The program takes input messages from the client as capital letters
representing products, handles exceptions of wrong inputs, and then
displays the current total sum after each introduced item.
When the customer presses Enter to pay, the transaction process finishes and the
final price is displayed.

Signed-off-by: mihai.state <mihaita.state@yahoo.com>